--- frontend/src/Editor/LeftSidebar/SidebarPageSelector.jsx
@@ -1,618 +1,0 @@
-import React, { useState } from 'react';
-import usePinnedPopover from '@/_hooks/usePinnedPopover';
-import { LeftSidebarItem } from './SidebarItem';
-import { SidebarPinnedButton } from './SidebarPinnedButton';
-import _ from 'lodash';
-import { OverlayTrigger, Popover, Modal } from 'react-bootstrap';
-import Fuse from 'fuse.js';
-import { Button, HeaderSection } from '@/_ui/LeftSidebar';
-import { Alert } from '@/_ui/Alert';
-import { ToolTip } from '@/Editor/Inspector/Elements/Components/ToolTip';
-
-export const LeftSidebarPageSelector = ({
-  darkMode,
-  appDefinition,
-  currentPageId,
-  addNewPage,
-  switchPage,
-  deletePage,
-  renamePage,
-  updateHomePage,
-  updatePageHandle,
-  queryPanelHeight,
-}) => {
-  const [open, trigger, content, popoverPinned, updatePopoverPinnedState] = usePinnedPopover(false);
-
-  const handlePopoverPinnedState = () => {
-    if (!popoverPinned) {
-      updatePopoverPinnedState(true);
-    }
-  };
-
-  const pages = Object.entries(appDefinition.pages).map(([id, page]) => ({ id, ...page }));
-  const [allpages, setPages] = useState(pages);
-
-  const { isExpanded } = JSON.parse(localStorage.getItem('queryManagerPreferences'));
-  const pageSelectorHeight = !isExpanded ? window.innerHeight - 85 : (queryPanelHeight * window.innerHeight) / 100 - 45;
-  const isHomePage = appDefinition.homePageId === currentPageId;
-
-  const [newPageBeingCreated, setNewPageBeingCreated] = useState(false);
-
-  const filterPages = (value) => {
-    if (!value || value.length === 0) return clearSearch();
-
-    const fuse = new Fuse(allpages, { keys: ['name'], threshold: 0.3 });
-    const result = fuse.search(value);
-
-    setPages(result.map((item) => item.item));
-  };
-
-  const clearSearch = () => {
-    setPages(pages);
-  };
-
-  React.useEffect(() => {
-    if (!_.isEqual(pages, allpages)) {
-      setPages(pages);
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [JSON.stringify(pages)]);
-
-  return (
-    <>
-      <LeftSidebarItem
-        tip="Pages"
-        {...trigger}
-        icon="page"
-        className={`left-sidebar-item left-sidebar-layout ${open && 'active'} left-sidebar-page-selector`}
-        text={'Pages'}
-      />
-      <div
-        {...content}
-        className={`card popover left-sidebar-page-selector ${open || popoverPinned ? 'show' : 'hide'} ${
-          darkMode && 'dark'
-        } `}
-        style={{
-          minWidth: '295px',
-          top: '45px',
-          borderRadius: '0px',
-          height: pageSelectorHeight,
-          maxHeight: window.innerHeight,
-          overflowX: 'hidden',
-        }}
-      >
-        <div className="card-body p-0" onClick={(event) => event.stopPropagation()}>
-          <HeaderSection darkMode={darkMode}>
-            <HeaderSection.PanelHeader title="Pages">
-              <div className="d-flex justify-content-end">
-                <Button
-                  onClick={() => setNewPageBeingCreated(true)}
-                  darkMode={darkMode}
-                  size="sm"
-                  styles={{ width: '76px' }}
-                >
-                  <Button.Content title={'Add'} iconSrc={'assets/images/icons/plus.svg'} direction="left" />
-                </Button>
-
-                <Button
-                  darkMode={darkMode}
-                  onClick={null} //Todo: global page settings
-                  size="sm"
-                  styles={{ width: '28px', padding: 0 }}
-                >
-                  <Button.Content iconSrc="assets/images/icons/editor/left-sidebar/settings.svg" />
-                </Button>
-
-                <SidebarPinnedButton
-                  darkMode={darkMode}
-                  component={'PageSelector'}
-                  state={popoverPinned}
-                  updateState={updatePopoverPinnedState}
-                />
-              </div>
-            </HeaderSection.PanelHeader>
-            <HeaderSection.SearchBoxComponent onChange={filterPages} placeholder={'Search'} placeholderIcon={'⌘S'} />
-          </HeaderSection>
-
-          <div className={`${darkMode && 'dark'} page-selector-panel-body`}>
-            <div className="list-group">
-              {allpages.map((page) => (
-                <div key={page.id} className="page-handler">
-                  <PageHandler
-                    darkMode={darkMode}
-                    page={page}
-                    isSelected={page.id === currentPageId}
-                    switchPage={switchPage}
-                    deletePage={deletePage}
-                    renamePage={renamePage}
-                    updatePopoverPinnedState={handlePopoverPinnedState}
-                    isHomePage={isHomePage}
-                    updateHomePage={updateHomePage}
-                    updatePageHandle={updatePageHandle}
-                  />
-                </div>
-              ))}
-              {newPageBeingCreated && (
-                <div className="page-handler">
-                  <AddingPageHandler
-                    addNewPage={addNewPage}
-                    setNewPageBeingCreated={setNewPageBeingCreated}
-                    switchPage={switchPage}
-                  />
-                </div>
-              )}
-            </div>
-          </div>
-        </div>
-      </div>
-    </>
-  );
-};
-
-const PageHandler = ({
-  darkMode,
-  page,
-  isSelected,
-  switchPage,
-  deletePage,
-  renamePage,
-  updatePopoverPinnedState,
-  isHomePage,
-  updateHomePage,
-  updatePageHandle,
-}) => {
-  const [isEditingPageName, setIsEditingPageName] = useState(false);
-
-  const [showEditModal, setShowEditModal] = useState(false);
-  const [showPagehandlerMenu, setShowPagehandlerMenu] = useState(false);
-
-  const handleClose = () => {
-    setShowEditModal(false);
-    setShowPagehandlerMenu(true);
-  };
-  const handleShow = () => {
-    setShowEditModal(true);
-    setShowPagehandlerMenu(false);
-  };
-
-  const handleCallback = (id) => {
-    switch (id) {
-      case 'delete-page':
-        deletePage(page.id, isHomePage);
-        break;
-
-      case 'rename-page':
-        updatePopoverPinnedState(true);
-        setIsEditingPageName(true);
-        break;
-
-      case 'mark-as-home-page':
-        updateHomePage(page.id);
-        break;
-
-      case 'edit-page-handle':
-        handleShow();
-        updatePopoverPinnedState(true);
-        break;
-
-      default:
-        break;
-    }
-  };
-
-  if (isEditingPageName) {
-    return <RenameInput page={page} updaterCallback={renamePage} updatePageEditMode={setIsEditingPageName} />;
-  }
-  const windowUrl = window.location.href;
-
-  const slug = windowUrl.split(page.handle)[0];
-  return (
-    <div
-      className={`card cursor-pointer ${isSelected ? 'active' : 'non-active-page'}`}
-      onClick={() => switchPage(page.id)}
-    >
-      <div className="card-body">
-        <div className="row" role="button">
-          <div className="col-auto">
-            {isSelected && (
-              <svg width="8" height="14" viewBox="0 0 8 14" fill="none" xmlns="http://www.w3.org/2000/svg">
-                <path
-                  d="M0.666667 1.66667C0.666667 2.03486 0.965143 2.33333 1.33333 2.33333C1.70152 2.33333 2 2.03486 2 1.66667C2 1.29848 1.70152 1 1.33333 1C0.965143 1 0.666667 1.29848 0.666667 1.66667Z"
-                  stroke="#8092AC"
-                  strokeWidth="1.33333"
-                />
-                <path
-                  d="M5.99992 1.66667C5.99992 2.03486 6.2984 2.33333 6.66659 2.33333C7.03478 2.33333 7.33325 2.03486 7.33325 1.66667C7.33325 1.29848 7.03478 1 6.66659 1C6.2984 1 5.99992 1.29848 5.99992 1.66667Z"
-                  stroke="#8092AC"
-                  strokeWidth="1.33333"
-                />
-                <path
-                  d="M0.666667 7.00001C0.666667 7.3682 0.965143 7.66668 1.33333 7.66668C1.70152 7.66668 2 7.3682 2 7.00001C2 6.63182 1.70152 6.33334 1.33333 6.33334C0.965143 6.33334 0.666667 6.63182 0.666667 7.00001Z"
-                  stroke="#8092AC"
-                  strokeWidth="1.33333"
-                />
-                <path
-                  d="M5.99992 7.00001C5.99992 7.3682 6.2984 7.66668 6.66659 7.66668C7.03478 7.66668 7.33325 7.3682 7.33325 7.00001C7.33325 6.63182 7.03478 6.33334 6.66659 6.33334C6.2984 6.33334 5.99992 6.63182 5.99992 7.00001Z"
-                  stroke="#8092AC"
-                  strokeWidth="1.33333"
-                />
-                <path
-                  d="M0.666667 12.3333C0.666667 12.7015 0.965143 13 1.33333 13C1.70152 13 2 12.7015 2 12.3333C2 11.9651 1.70152 11.6667 1.33333 11.6667C0.965143 11.6667 0.666667 11.9651 0.666667 12.3333Z"
-                  stroke="#8092AC"
-                  strokeWidth="1.33333"
-                />
-                <path
-                  d="M5.99992 12.3333C5.99992 12.7015 6.2984 13 6.66659 13C7.03478 13 7.33325 12.7015 7.33325 12.3333C7.33325 11.9651 7.03478 11.6667 6.66659 11.6667C6.2984 11.6667 5.99992 11.9651 5.99992 12.3333Z"
-                  stroke="#8092AC"
-                  strokeWidth="1.33333"
-                />
-              </svg>
-            )}
-          </div>
-          <div className="col text-truncate" data-cy="event-handler">
-            {page.name}
-          </div>
-          <div className="col-auto">
-            {isSelected && isHomePage && <img src="assets/images/icons/home.svg" height={14} width={14} />}
-          </div>
-          <div className="col-auto">
-            {isSelected && (
-              <PagehandlerMenu
-                page={page}
-                darkMode={darkMode}
-                handlePageCallback={handleCallback}
-                showMenu={showPagehandlerMenu}
-                setShowMenu={setShowPagehandlerMenu}
-              />
-            )}
-            <EditModal
-              slug={slug}
-              page={page}
-              show={showEditModal}
-              handleClose={handleClose}
-              updatePageHandle={updatePageHandle}
-              darkMode={darkMode}
-            />
-          </div>
-        </div>
-      </div>
-    </div>
-  );
-};
-
-const AddingPageHandler = ({ addNewPage, setNewPageBeingCreated }) => {
-  const handleAddingNewPage = (pageName) => {
-    if (pageName) {
-      addNewPage({ name: pageName, handle: _.kebabCase(pageName.toLowerCase()) });
-    }
-    setNewPageBeingCreated(false);
-  };
-
-  return (
-    <div className="row" role="button">
-      <div className="col-12">
-        <input
-          type="text"
-          className="form-control page-name-input"
-          autoFocus
-          onBlur={(event) => {
-            const name = event.target.value;
-            handleAddingNewPage(name);
-            event.stopPropagation();
-          }}
-          onKeyDown={(event) => {
-            if (event.key === 'Enter') {
-              const name = event.target.value;
-              handleAddingNewPage(name);
-              event.stopPropagation();
-            }
-          }}
-        />
-      </div>
-    </div>
-  );
-};
-
-const RenameInput = ({ page, updaterCallback, updatePageEditMode }) => {
-  const handleAddingNewPage = (pageName) => {
-    if (pageName.length > 0 && pageName !== page.name) {
-      updaterCallback(page.id, pageName);
-    }
-    updatePageEditMode(false);
-  };
-
-  return (
-    <div className="row" role="button">
-      <div className="col-12">
-        <input
-          type="text"
-          className="form-control page-name-input"
-          autoFocus
-          defaultValue={page.name}
-          onBlur={(event) => {
-            const name = event.target.value;
-            handleAddingNewPage(name);
-            event.stopPropagation();
-          }}
-          onKeyDown={(event) => {
-            if (event.key === 'Enter') {
-              const name = event.target.value;
-              handleAddingNewPage(name);
-              event.stopPropagation();
-            }
-          }}
-        />
-      </div>
-    </div>
-  );
-};
-
-const PagehandlerMenu = ({ page, darkMode, handlePageCallback, showMenu, setShowMenu }) => {
-  const closeMenu = () => {
-    setShowMenu(false);
-  };
-
-  React.useEffect(() => {
-    const handleClickOutside = (event) => {
-      if (showMenu && event.target.closest('.pagehandler-menu') === null) {
-        closeMenu();
-      }
-    };
-
-    document.addEventListener('mousedown', handleClickOutside);
-    return () => {
-      document.removeEventListener('mousedown', handleClickOutside);
-    };
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [JSON.stringify({ page, showMenu })]);
-
-  return (
-    <OverlayTrigger
-      trigger={'click'}
-      placement={'bottom'}
-      rootClose
-      show={showMenu}
-      overlay={
-        <Popover key={page.id} id="page-handler-menu" className={darkMode && 'popover-dark-themed'}>
-          <Popover.Content key={page.id} bsPrefix="popover-body">
-            <div className="card-body">
-              <PageHandleField page={page} updatePageHandle={handlePageCallback} />
-              <hr style={{ margin: '0.75rem 0' }} />
-              <div className="menu-options mb-0">
-                <Field
-                  id="rename-page"
-                  text="Rename"
-                  iconSrc={'assets/images/icons/input.svg'}
-                  closeMenu={closeMenu}
-                  callback={handlePageCallback}
-                />
-                <Field
-                  id="duplicate-page"
-                  text="Duplicate"
-                  iconSrc={'assets/images/icons/duplicate.svg'}
-                  closeMenu={closeMenu}
-                  callback={handlePageCallback}
-                />
-                <Field
-                  id="mark-as-home-page"
-                  text="Mark home"
-                  iconSrc={'assets/images/icons/home.svg'}
-                  closeMenu={closeMenu}
-                  callback={handlePageCallback}
-                />
-
-                <Field
-                  id="hide-page"
-                  text="Hide Page"
-                  iconSrc={'assets/images/icons/eye.svg'}
-                  closeMenu={closeMenu}
-                  callback={handlePageCallback}
-                />
-
-                <Field
-                  id="delete-page"
-                  text="Delete page"
-                  iconSrc={'assets/images/icons/delete.svg'}
-                  customClass="field__danger"
-                  closeMenu={closeMenu}
-                  callback={handlePageCallback}
-                />
-              </div>
-            </div>
-          </Popover.Content>
-        </Popover>
-      }
-    >
-      <Button.UnstyledButton onClick={() => setShowMenu(true)} styles={{ height: '20px' }}>
-        <Button.Content iconSrc={'assets/images/icons/3dots-menu.svg'} />
-      </Button.UnstyledButton>
-    </OverlayTrigger>
-  );
-};
-
-const Field = ({ id, text, iconSrc, customClass = '', closeMenu, callback = () => null }) => {
-  const handleOnClick = () => {
-    closeMenu();
-    callback(id);
-  };
-
-  return (
-    <div className={`field ${customClass ? ` ${customClass}` : ''}`}>
-      <Button.UnstyledButton onClick={handleOnClick} styles={{ height: '28px' }}>
-        <Button.Content title={text} iconSrc={iconSrc} direction="left" />
-      </Button.UnstyledButton>
-    </div>
-  );
-};
-
-const PageHandleField = ({ page, updatePageHandle }) => {
-  const Label = () => {
-    return (
-      <label htmlFor="pin" className="form-label">
-        Page Handle
-      </label>
-    );
-  };
-
-  const content = () => {
-    return (
-      <div className="col">
-        <span style={{ color: '#889096' }}>.../</span>
-        <span>{page.handle}</span>
-      </div>
-    );
-  };
-
-  return (
-    <div className="mb-2 px-2">
-      <Label />
-      <Button.UnstyledButton
-        onClick={() => {
-          updatePageHandle('edit-page-handle');
-        }}
-        classNames="page-handle-button-container"
-      >
-        <Button.Content title={content} iconSrc={'assets/images/icons/input.svg'} direction="right" />
-      </Button.UnstyledButton>
-    </div>
-  );
-};
-
-const EditModal = ({ slug, page, show, handleClose, updatePageHandle, darkMode }) => {
-  const [pageHandle, setPageHandle] = useState(page.handle);
-  const [error, setError] = useState(null);
-  const [isSaving, setIsSaving] = useState(false);
-
-  React.useEffect(() => {
-    setError(null);
-  }, [show]);
-
-  const handleSave = () => {
-    if (pageHandle === page.handle) {
-      return handleClose();
-    }
-
-    setIsSaving(true);
-    updatePageHandle(page.id, pageHandle);
-    setTimeout(() => {
-      setIsSaving(false);
-      return handleClose();
-    }, 900);
-  };
-
-  const handleCancel = () => {
-    setError(null);
-    setIsSaving(false);
-    handleClose();
-  };
-
-  return (
-    <Modal
-      show={show}
-      onHide={handleClose}
-      size="sm"
-      centered
-      className={`${darkMode && 'theme-dark'} page-handle-edit-modal `}
-    >
-      <Modal.Header>
-        <Modal.Title style={{ fontSize: '16px', fontWeight: '400' }}>Edit page handle</Modal.Title>
-        <span className="cursor-pointer" size="sm" onClick={() => handleClose()}>
-          <svg
-            xmlns="http://www.w3.org/2000/svg"
-            className="icon icon-tabler icon-tabler-x"
-            width="24"
-            height="24"
-            viewBox="0 0 24 24"
-            strokeWidth="2"
-            stroke="currentColor"
-            fill="none"
-            strokeLinecap="round"
-            strokeLinejoin="round"
-          >
-            <path stroke="none" d="M0 0h24v24H0z" fill="none"></path>
-            <line x1="18" y1="6" x2="6" y2="18"></line>
-            <line x1="6" y1="6" x2="18" y2="18"></line>
-          </svg>
-        </span>
-      </Modal.Header>
-      <Modal.Body>
-        <div className="page-handle-edit-container mb-4">
-          <EditInput
-            slug={slug}
-            error={error}
-            setError={setError}
-            pageHandle={pageHandle}
-            setPageHandle={setPageHandle}
-            isSaving={isSaving}
-          />
-        </div>
-
-        <div className="alert-container">
-          <Alert svg="alert-info" cls="page-handler-alert">
-            some warning about the effects changing pagehandle will create.
-          </Alert>
-        </div>
-      </Modal.Body>
-      <Modal.Footer>
-        <Button darkMode={darkMode} onClick={handleCancel} styles={{ height: '32px' }} disabled={isSaving}>
-          <Button.Content title="Cancel" />
-        </Button>
-        <Button
-          darkMode={darkMode}
-          onClick={handleSave}
-          styles={{ backgroundColor: '#3E63DD', color: '#FDFDFE', height: '32px' }}
-          disabled={error !== null}
-          isLoading={isSaving}
-        >
-          <Button.Content title="Save" iconSrc="assets/images/icons/save.svg" direction="left" />
-        </Button>
-      </Modal.Footer>
-    </Modal>
-  );
-};
-
-const EditInput = ({ slug, error, setError, pageHandle, setPageHandle, isSaving = false }) => {
-  const [value, set] = useState(pageHandle);
-
-  const onChangePageHandleValue = (event) => {
-    setError(null);
-    const newHandle = event.target.value;
-
-    if (newHandle === '') setError('Page handle cannot be empty');
-    if (newHandle === value) setError('Page handle cannot be same as the existing page handle');
-
-    set(newHandle);
-  };
-
-  React.useEffect(() => {
-    if (!isSaving) {
-      setPageHandle(value);
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [value]);
-
-  React.useEffect(() => {
-    if (isSaving) {
-      set(pageHandle);
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [isSaving]);
-
-  const label = <span style={{ color: '#889096' }}>{slug.substring(0, 16)}.../</span>;
-
-  return (
-    <div className="input-group col">
-      <div className="input-group-text">
-        <ToolTip label={label} meta={{ tip: slug }} labelClass="page-handle-tip" />
-      </div>
-      <input
-        type="text"
-        className={`page-handler-input form-control form-control-sm ${error ? 'is-invalid' : ''}`}
-        placeholder={'Enter page handle'}
-        onChange={onChangePageHandleValue}
-        value={pageHandle}
-      />
-      <div className="invalid-feedback">{error}</div>
-    </div>
-  );
-};
