+++ frontend/src/Editor/LeftSidebar/SidebarPageSelector/EditModal.jsx
@@ -1,0 +1,98 @@
+import React, { useState } from 'react';
+import { Modal } from 'react-bootstrap';
+import { Button } from '@/_ui/LeftSidebar';
+import { Alert } from '@/_ui/Alert';
+import { EditInput } from './EditInput';
+
+export const EditModal = ({ slug, page, show, handleClose, updatePageHandle, darkMode }) => {
+  const [pageHandle, setPageHandle] = useState(page.handle);
+  const [error, setError] = useState(null);
+  const [isSaving, setIsSaving] = useState(false);
+
+  React.useEffect(() => {
+    setError(null);
+  }, [show]);
+
+  const handleSave = () => {
+    if (pageHandle === page.handle) {
+      return handleClose();
+    }
+
+    setIsSaving(true);
+    updatePageHandle(page.id, pageHandle);
+    setTimeout(() => {
+      setIsSaving(false);
+      return handleClose();
+    }, 900);
+  };
+
+  const handleCancel = () => {
+    setError(null);
+    setIsSaving(false);
+    handleClose();
+  };
+
+  return (
+    <Modal
+      show={show}
+      onHide={handleClose}
+      size="sm"
+      centered
+      className={`${darkMode && 'theme-dark'} page-handle-edit-modal `}
+    >
+      <Modal.Header>
+        <Modal.Title style={{ fontSize: '16px', fontWeight: '400' }}>Edit page handle</Modal.Title>
+        <span className="cursor-pointer" size="sm" onClick={() => handleClose()}>
+          <svg
+            xmlns="http://www.w3.org/2000/svg"
+            className="icon icon-tabler icon-tabler-x"
+            width="24"
+            height="24"
+            viewBox="0 0 24 24"
+            strokeWidth="2"
+            stroke="currentColor"
+            fill="none"
+            strokeLinecap="round"
+            strokeLinejoin="round"
+          >
+            <path stroke="none" d="M0 0h24v24H0z" fill="none"></path>
+            <line x1="18" y1="6" x2="6" y2="18"></line>
+            <line x1="6" y1="6" x2="18" y2="18"></line>
+          </svg>
+        </span>
+      </Modal.Header>
+      <Modal.Body>
+        <div className="page-handle-edit-container mb-4">
+          <EditInput
+            slug={slug}
+            error={error}
+            setError={setError}
+            pageHandle={pageHandle}
+            setPageHandle={setPageHandle}
+            isSaving={isSaving}
+          />
+        </div>
+
+        <div className="alert-container">
+          <Alert svg="alert-info" cls="page-handler-alert">
+            some warning about the effects changing pagehandle will create.
+          </Alert>
+        </div>
+      </Modal.Body>
+      <Modal.Footer>
+        <Button darkMode={darkMode} onClick={handleCancel} styles={{ height: '32px' }} disabled={isSaving}>
+          <Button.Content title="Cancel" />
+        </Button>
+        <Button
+          darkMode={darkMode}
+          onClick={handleSave}
+          styles={{ backgroundColor: '#3E63DD', color: '#FDFDFE', height: '32px' }}
+          disabled={error !== null}
+          isLoading={isSaving}
+        >
+          <Button.Content title="Save" iconSrc="assets/images/icons/save.svg" direction="left" />
+        </Button>
+      </Modal.Footer>
+    </Modal>
+  );
+};
